requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
statsmodels>=0.14.0
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
plotly>=5.15.0

# Machine Learning
xgboost>=1.7.0
lightgbm>=4.0.0
torch>=2.0.0
pytorch-lightning>=2.0.0

# Time Series
pmdarima>=2.0.0
prophet>=1.1.0
darts>=0.24.0
tsfresh>=0.20.0

# Portfolio & Risk
cvxpy>=1.3.0
riskfolio-lib>=4.0.0
PyPortfolioOpt>=1.5.0
arch>=6.0.0

# Backtesting
vectorbt>=0.25.0

# Data sources
yfinance>=0.2.0
alpha-vantage>=2.3.0

# Web framework for demo
streamlit>=1.25.0

# Development
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Tracking
mlflow>=2.5.0
wandb>=0.15.0
